feat(ui): created step dir, moved add city step to dir
