+++ app/src/app/[lng]/onboarding/setup/page.tsx
@@ -1,10 +1,7 @@
 "use client";
 
-import RecentSearches from "@/components/recent-searches";
 import WizardSteps from "@/components/wizard-steps";
-import { set } from "@/features/city/openclimateCitySlice";
 import { useTranslation } from "@/i18n/client";
-import { useAppDispatch } from "@/lib/hooks";
 import type { CityAttributes } from "@/models/City";
 import {
   api,
@@ -26,7 +23,6 @@ import {
   ArrowForwardIcon,
   CheckIcon,
   InfoOutlineIcon,
-  SearchIcon,
   WarningIcon,
 } from "@chakra-ui/icons";
 import { Link } from "@chakra-ui/next-js";
@@ -34,7 +30,6 @@ import {
   Box,
   Button,
   Card,
-  Flex,
   FormControl,
   FormErrorIcon,
   FormErrorMessage,
@@ -46,10 +41,7 @@ import {
   InputGroup,
   InputLeftElement,
   InputRightElement,
-  Radio,
-  RadioGroup,
   Select,
-  Stack,
   Text,
   useOutsideClick,
   useSteps,
@@ -62,7 +54,7 @@ import {
 import type { TFunction } from "i18next";
 import dynamic from "next/dynamic";
 import { useRouter } from "next/navigation";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import type {
   Control,
   FieldErrors,
@@ -73,17 +65,18 @@ import { Controller, useForm } from "react-hook-form";
 import { Trans } from "react-i18next/TransWithoutContext";
 import { MdOutlineAspectRatio, MdOutlinePeopleAlt } from "react-icons/md";
 import FormattedThousandsNumberInput from "@/app/[lng]/onboarding/setup/FormattedThousandsNumberInput";
-import Image from "next/image";
+
 import {
   CalenderIcon,
   DataFormatIcon,
   InventoryButtonCheckIcon,
 } from "@/components/icons";
 import { CircleFlag } from "react-circle-flags";
+import SelectCityStep from "../steps/select-city-steps";
 
 const CityMap = dynamic(() => import("@/components/CityMap"), { ssr: false });
 
-type Inputs = {
+export type Inputs = {
   city: string;
   year: number;
   inventoryGoal: string;
@@ -97,635 +90,16 @@ type Inputs = {
   totalCountryEmissions: number;
 };
 
-type PopulationEntry = {
-  year: number;
-  population: number;
-  datasource_id: string;
-};
-
-type CountryEmissionsEntry = {
+export type CountryEmissionsEntry = {
   year: number;
   total_emissions: number;
 };
 
-type OnboardingData = {
+export type OnboardingData = {
   name: string;
   locode: string;
   year: number;
 };
-
-function SelectCityStep({
-  errors,
-  register,
-  control,
-  t,
-  setValue,
-  watch,
-  ocCityData,
-  setOcCityData,
-  setData,
-}: {
-  errors: FieldErrors<Inputs>;
-  register: UseFormRegister<Inputs>;
-  control: Control<Inputs>;
-  t: TFunction;
-  setValue: any;
-  watch: Function;
-  ocCityData?: OCCityAttributes;
-  setOcCityData: (cityData: OCCityAttributes) => void;
-  setData: (data: OnboardingData) => void;
-}) {
-  const currentYear = new Date().getFullYear();
-  const numberOfYearsDisplayed = 10;
-  const years = Array.from(
-    { length: numberOfYearsDisplayed },
-    (_x, i) => currentYear - i,
-  );
-
-  const dispatch = useAppDispatch();
-
-  const [onInputClicked, setOnInputClicked] = useState<boolean>(false);
-  const [isCityNew, setIsCityNew] = useState<boolean>(false);
-  const [locode, setLocode] = useState<string | null>(null);
-
-  const yearInput = watch("year");
-  const year: number | null = yearInput ? parseInt(yearInput) : null;
-  const cityInputQuery = watch("city");
-  const cityPopulationYear = watch("cityPopulationYear");
-  const regionPopulationYear = watch("regionPopulationYear");
-  const countryPopulationYear = watch("countryPopulationYear");
-
-  const handleSetCity = (city: OCCityAttributes) => {
-    console.log("city", city);
-    setValue("city", city.name);
-    setOnInputClicked(false);
-    dispatch(set(city));
-    setLocode(city.actor_id);
-    setOcCityData(city);
-
-    if (year) {
-      setData({
-        name: city.name,
-        locode: city.actor_id!,
-        year: year!,
-      });
-    }
-
-    setIsCityNew(true);
-  };
-
-  console.log(locode);
-
-  useEffect(() => {
-    if (year && ocCityData) {
-      setData({
-        name: ocCityData.name,
-        locode: ocCityData.actor_id,
-        year: year!,
-      });
-    }
-  }, [year, ocCityData, setData]);
-
-  useEffect(() => {
-    if (!cityInputQuery || cityInputQuery.length === 0) {
-      setOnInputClicked(false);
-      setIsCityNew(false);
-    }
-  }, [cityInputQuery]);
-
-  const { data: cityData } = useGetOCCityDataQuery(locode!, {
-    skip: !locode,
-  });
-  const countryLocode =
-    locode && locode.length > 0 ? locode.split(" ")[0] : null;
-  const { data: countryData } = useGetOCCityDataQuery(countryLocode!, {
-    skip: !countryLocode,
-  });
-  const regionLocode = cityData?.is_part_of;
-  const { data: regionData } = useGetOCCityDataQuery(regionLocode!, {
-    skip: !regionLocode,
-  });
-
-  // react to API data changes and different year selections
-  useEffect(() => {
-    if (cityData && year) {
-      const population = findClosestYear(
-        cityData.population,
-        year,
-        numberOfYearsDisplayed,
-      );
-      if (!population) {
-        console.error("Failed to find population data for city");
-        return;
-      }
-      setValue("cityPopulation", population.population);
-      setValue("cityPopulationYear", population.year);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [cityData, year, setValue]);
-
-  useEffect(() => {
-    if (regionData && year) {
-      const population = findClosestYear(
-        regionData.population,
-        year,
-        numberOfYearsDisplayed,
-      );
-      if (!population) {
-        console.error("Failed to find population data for region");
-        return;
-      }
-      setValue("regionPopulation", population.population);
-      setValue("regionPopulationYear", population.year);
-    }
-  }, [regionData, year, setValue]);
-
-  useEffect(() => {
-    if (countryData && year) {
-      const population = findClosestYear(
-        countryData.population,
-        year,
-        numberOfYearsDisplayed,
-      );
-      if (!population) {
-        console.error("Failed to find population data for region");
-        return;
-      }
-      setValue("countryPopulation", population.population);
-      setValue("countryPopulationYear", population.year);
-      const keys = Object.keys(countryData.emissions);
-      const sourceId = keys.find((id) => id.startsWith("UNFCCC"));
-
-      if (sourceId) {
-        const emissionsData: CountryEmissionsEntry[] =
-          countryData.emissions[sourceId].data;
-        const emissions = emissionsData.find(
-          (e) => e.year === year,
-        )?.total_emissions;
-        if (emissions == null) {
-          console.error("Failed to find country emissions for ", year);
-        }
-        setValue("totalCountryEmissions", emissions);
-      }
-    }
-  }, [countryData, year, setValue]);
-
-  // import custom redux hooks
-  const {
-    data: cities,
-    isLoading,
-    isSuccess,
-  } = useGetOCCityQuery(cityInputQuery, {
-    skip: cityInputQuery?.length <= 2,
-  });
-
-  const renderParentPath = (path: []) => {
-    let pathString = "";
-    const pathCopy = [...path];
-
-    pathCopy
-      ?.reverse()
-      .slice(1)
-      .map((parent: any) => {
-        if (pathString) {
-          pathString = pathString + " > ";
-        }
-        pathString = pathString + parent.name;
-      });
-
-    return pathString;
-  };
-
-  // using useOutsideClick instead of onBlur input attribute
-  // to fix clicking city dropdown entries not working
-  const cityInputRef = useRef<HTMLDivElement>(null);
-  useOutsideClick({
-    ref: cityInputRef,
-    handler: () => setTimeout(() => setOnInputClicked(false), 0),
-  });
-
-  return (
-    <Box w="full">
-      <Box
-        minW={400}
-        w="full"
-        display="flex"
-        flexDir="column"
-        gap="24px"
-        mb="48px"
-      >
-        <Heading size="xl">{t("setup-city-heading")}</Heading>
-        <Text
-          color="content.tertiary"
-          fontSize="body.lg"
-          fontStyle="normal"
-          fontWeight="400"
-          letterSpacing="wide"
-        >
-          {t("setup-city-details")}
-        </Text>
-      </Box>
-      <Box w="full">
-        <Card p={6} shadow="none" px="24px" py="32px">
-          <form className="space-y-8">
-            <FormControl isInvalid={!!errors.city}>
-              <FormLabel>{t("city")}</FormLabel>
-              <InputGroup
-                ref={cityInputRef}
-                shadow="1dp"
-                bg={errors.city ? "sentiment.negativeOverlay" : "base.light"}
-                borderRadius="8px"
-              >
-                <InputLeftElement pointerEvents="none" borderRadius="none">
-                  <SearchIcon color="tertiary" boxSize={4} mt={2} ml={4} />
-                </InputLeftElement>
-                <Input
-                  type="text"
-                  placeholder={t("select-city-placeholder")}
-                  size="lg"
-                  {...register("city", {
-                    required: t("select-city-required"),
-                  })}
-                  autoComplete="off"
-                  onKeyUp={() => setOnInputClicked(true)}
-                  onFocus={() => setOnInputClicked(true)}
-                />
-                <InputRightElement>
-                  {isCityNew && (
-                    <CheckIcon
-                      color="semantic.success"
-                      boxSize={4}
-                      mr={4}
-                      mt={2}
-                    />
-                  )}
-                </InputRightElement>
-              </InputGroup>
-              {onInputClicked && (
-                <Box
-                  shadow="2dp"
-                  className="h-auto max-h-[272px] transition-all duration-150 overflow-scroll flex flex-col py-3 gap-3 rounded-lg w-full absolute bg-white z-50 mt-2 border border-[1px solid #E6E7FF]"
-                >
-                  {!isLoading && !cityInputQuery && <RecentSearches />}
-                  {isLoading && <p className="px-4">Fetching Cities...</p>}
-                  {isSuccess &&
-                    cities &&
-                    cities.map((city: OCCityAttributes) => {
-                      return (
-                        <Box
-                          onClick={() => handleSetCity(city)}
-                          key={city.actor_id}
-                          className="h-[72px] py-3 w-full flex flex-col justify-center group px-4 hover:bg-[#2351DC] transition-all duration-150 cursor-pointer"
-                        >
-                          <Text
-                            className="group-hover:text-white"
-                            color="content.secondary"
-                            fontSize="body.lg"
-                            fontFamily="body"
-                            fontWeight="normal"
-                            lineHeight="24"
-                            letterSpacing="wide"
-                          >
-                            {city.name}
-                          </Text>
-                          <Text
-                            className="group-hover:text-[#E8EAFB]"
-                            color="content.tertiary"
-                            fontSize="body.lg"
-                            fontFamily="body.md"
-                            fontWeight="normal"
-                            lineHeight="20"
-                            letterSpacing="wide"
-                          >
-                            {renderParentPath(city.root_path_geo)}
-                          </Text>
-                        </Box>
-                      );
-                    })}
-                </Box>
-              )}
-              <FormErrorMessage gap="6px">
-                <WarningIcon />
-                <Text
-                  fontSize="body.md"
-                  color="content.tertiary"
-                  fontStyle="normal"
-                >
-                  {errors.city && errors.city.message}
-                </Text>
-              </FormErrorMessage>
-            </FormControl>
-            {ocCityData ? (
-              <Box
-                display="flex"
-                flexDir="column"
-                borderRadius="8px"
-                w="full"
-                h="full"
-                gap="24px"
-                overflow="hidden"
-              >
-                <CityMap
-                  locode={ocCityData.actor_id}
-                  height={500}
-                  width={1100}
-                />
-                <Box display="flex" alignItems="center" gap="6px">
-                  <InfoOutlineIcon />
-                  <Text
-                    color="content.secondary"
-                    fontWeight="normal"
-                    letterSpacing="wide"
-                  >
-                    In case the geographical boundary is not the right one{" "}
-                    <Link href="">
-                      <Text
-                        as="span"
-                        color="interactive.secondary"
-                        fontWeight="600"
-                        letterSpacing="wide"
-                        textDecorationLine="underline"
-                      >
-                        Contact Us
-                      </Text>
-                    </Link>
-                  </Text>
-                </Box>
-              </Box>
-            ) : (
-              <Box
-                bg="base.light"
-                h="317px"
-                w="full"
-                display="flex"
-                alignItems="center"
-                justifyContent="center"
-                flexDir="column"
-                gap="24px"
-                borderWidth={1}
-                borderColor="border.neutral"
-                borderStyle="dashed"
-                borderRadius="8px"
-              >
-                <Image
-                  src="/assets/city-image.svg"
-                  alt="city-image"
-                  height={400}
-                  width={200}
-                />
-                <Box display="flex" flexDir="column" gap="8px">
-                  <Text
-                    color="content.tertiary"
-                    fontSize="title.md"
-                    fontWeight="bold"
-                    lineHeight="24"
-                    fontFamily="heading"
-                    textAlign="center"
-                  >
-                    Search and select the city to be shown on the map
-                  </Text>
-                  <Text
-                    color="interactive.control"
-                    fontSize="body.md"
-                    fontWeight="400"
-                    fontStyle="normal"
-                    lineHeight="24"
-                    textAlign="center"
-                    letterSpacing="wide"
-                  >
-                    You will be able to check the geographical boundary for your
-                    inventory
-                  </Text>
-                </Box>
-              </Box>
-            )}
-
-            {/* <FormControl isInvalid={!!errors.year}>
-              <FormLabel>{t("inventory-year")}</FormLabel>
-              <InputGroup>
-                <Select
-                  placeholder={t("inventory-year-placeholder")}
-                  size="lg"
-                  {...register("year", {
-                    required: t("inventory-year-required"),
-                  })}
-                >
-                  {years.map((year: number, i: number) => (
-                    <option value={year} key={i}>
-                      {year}
-                    </option>
-                  ))}
-                </Select>
-                <InputRightElement>
-                  {!!year && (
-                    <CheckIcon
-                      color="semantic.success"
-                      boxSize={4}
-                      mt={2}
-                      mr={10}
-                    />
-                  )}
-                </InputRightElement>
-              </InputGroup>
-              <FormErrorMessage>
-                {errors.year && errors.year.message}
-              </FormErrorMessage>
-            </FormControl>
-            <HStack spacing={6} align="start">
-              <FormControl isInvalid={!!errors.cityPopulation}>
-                <FormLabel>{t("city-population-title")}</FormLabel>
-                <FormattedThousandsNumberInput<Inputs>
-                  name="cityPopulation"
-                  control={control}
-                  rules={{
-                    required: t("population-required"),
-                  }}
-                  placeholder={t("city-population-placeholder")}
-                  size="lg"
-                />
-                <FormErrorMessage
-                  color="content.tertiary"
-                  letterSpacing="0.5px"
-                >
-                  <FormErrorIcon />
-                  {errors.cityPopulation && errors.cityPopulation.message}
-                </FormErrorMessage>
-              </FormControl>
-
-              <FormControl isInvalid={!!errors.cityPopulationYear} w={60}>
-                <FormLabel>{t("population-year")}</FormLabel>
-                <InputGroup>
-                  <Select
-                    placeholder={t("year-placeholder")}
-                    size="lg"
-                    {...register("cityPopulationYear", {
-                      required: t("required"),
-                      valueAsNumber: true,
-                    })}
-                  >
-                    {years.map((year: number, i: number) => (
-                      <option value={year} key={i}>
-                        {year}
-                      </option>
-                    ))}
-                  </Select>
-                  {cityPopulationYear ? (
-                    <InputRightElement>
-                      <CheckIcon
-                        color="semantic.success"
-                        boxSize={4}
-                        mt={2}
-                        mr={10}
-                      />
-                    </InputRightElement>
-                  ) : null}
-                </InputGroup>
-                <FormErrorMessage
-                  color="content.tertiary"
-                  letterSpacing="0.5px"
-                >
-                  <FormErrorIcon />
-                  {errors.cityPopulationYear &&
-                    errors.cityPopulationYear.message}
-                </FormErrorMessage>
-              </FormControl>
-            </HStack>
-            <HStack spacing={6} align="start">
-              <FormControl isInvalid={!!errors.regionPopulation}>
-                <FormLabel>{t("region-population-title")}</FormLabel>
-                <FormattedThousandsNumberInput<Inputs>
-                  control={control}
-                  rules={{
-                    required: t("population-required"),
-                  }}
-                  placeholder={t("region-population-placeholder")}
-                  size="lg"
-                  {...register("regionPopulation", {
-                    required: t("population-required"),
-                    valueAsNumber: true,
-                  })}
-                  name="regionPopulation"
-                />
-                <FormErrorMessage
-                  color="content.tertiary"
-                  letterSpacing="0.5px"
-                >
-                  <FormErrorIcon />
-                  {errors.regionPopulation && errors.regionPopulation.message}
-                </FormErrorMessage>
-              </FormControl>
-              <FormControl isInvalid={!!errors.regionPopulationYear} w={60}>
-                <FormLabel>{t("population-year")}</FormLabel>
-                <InputGroup>
-                  <Select
-                    placeholder={t("year-placeholder")}
-                    size="lg"
-                    {...register("regionPopulationYear", {
-                      required: t("required"),
-                      valueAsNumber: true,
-                    })}
-                  >
-                    {years.map((year: number, i: number) => (
-                      <option value={year} key={i}>
-                        {year}
-                      </option>
-                    ))}
-                  </Select>
-                  {regionPopulationYear ? (
-                    <InputRightElement>
-                      <CheckIcon
-                        color="semantic.success"
-                        boxSize={4}
-                        mt={2}
-                        mr={10}
-                      />
-                    </InputRightElement>
-                  ) : null}
-                </InputGroup>
-                <FormErrorMessage
-                  color="content.tertiary"
-                  letterSpacing="0.5px"
-                >
-                  <FormErrorIcon />
-                  {errors.regionPopulationYear &&
-                    errors.regionPopulationYear.message}
-                </FormErrorMessage>
-              </FormControl>
-            </HStack>
-            <HStack spacing={6} align="start">
-              <FormControl isInvalid={!!errors.countryPopulation}>
-                <FormLabel>{t("country-population-title")}</FormLabel>
-                <FormattedThousandsNumberInput<Inputs>
-                  name="countryPopulation"
-                  control={control}
-                  rules={{
-                    required: t("population-required"),
-                  }}
-                  placeholder={t("country-population-placeholder")}
-                  size="lg"
-                />
-                <FormErrorMessage
-                  color="content.tertiary"
-                  letterSpacing="0.5px"
-                >
-                  <FormErrorIcon />
-                  {errors.countryPopulation && errors.countryPopulation.message}
-                </FormErrorMessage>
-              </FormControl>
-              <FormControl isInvalid={!!errors.countryPopulationYear} w={60}>
-                <FormLabel>{t("population-year")}</FormLabel>
-                <InputGroup>
-                  <Select
-                    placeholder={t("year-placeholder")}
-                    size="lg"
-                    {...register("countryPopulationYear", {
-                      required: t("required"),
-                      valueAsNumber: true,
-                    })}
-                  >
-                    {years.map((year: number, i: number) => (
-                      <option value={year} key={i}>
-                        {year}
-                      </option>
-                    ))}
-                  </Select>
-                  {countryPopulationYear ? (
-                    <InputRightElement>
-                      <CheckIcon
-                        color="semantic.success"
-                        boxSize={4}
-                        mt={2}
-                        mr={10}
-                      />
-                    </InputRightElement>
-                  ) : null}
-                </InputGroup>
-                <FormErrorMessage
-                  color="content.tertiary"
-                  letterSpacing="0.5px"
-                >
-                  <FormErrorIcon />
-                  {errors.countryPopulationYear &&
-                    errors.countryPopulationYear.message}
-                </FormErrorMessage>
-              </FormControl>
-            </HStack>
-            <HStack spacing={1.5} align="start">
-              <InfoOutlineIcon color="interactive.secondary" mt={1} />
-              <Text
-                color="content.tertiary"
-                fontSize="sm"
-                whiteSpace="pre-line"
-              >
-                {t("information-required")}
-              </Text>
-            </HStack> */}
-          </form>
-        </Card>
-      </Box>
-    </Box>
-  );
-}
 
 // Custom Radio Buttons
